move tests to feature_flags file
