+++ posthog/test/test_feature_flags.py
@@ -1,8 +1,10 @@
 import unittest
+from freezegun import freeze_time
 
 import mock
 
 from posthog.client import Client
+from posthog.request import APIError
 from posthog.test.test_utils import TEST_API_KEY
 
 
@@ -95,6 +97,254 @@ class TestLocalEvaluation(unittest.TestCase):
 
         self.assertTrue(feature_flag_match)
         self.assertFalse(not_feature_flag_match)
+    
+    @mock.patch("posthog.client.Poller")
+    @mock.patch("posthog.client.get")
+    def test_load_feature_flags(self, patch_get, patch_poll):
+        patch_get.return_value = {
+            "results": [
+                {"id": 1, "name": "Beta Feature", "key": "beta-feature", "active": True},
+                {"id": 2, "name": "Alpha Feature", "key": "alpha-feature", "active": False},
+            ]
+        }
+        client = Client(TEST_API_KEY, personal_api_key="test")
+        with freeze_time("2020-01-01T12:01:00.0000Z"):
+            client.load_feature_flags()
+        self.assertEqual(len(client.feature_flags), 1)
+        self.assertEqual(client.feature_flags[0]["key"], "beta-feature")
+        self.assertEqual(client._last_feature_flag_poll.isoformat(), "2020-01-01T12:01:00+00:00")
+        self.assertEqual(patch_poll.call_count, 1)
+
+    def test_load_feature_flags_wrong_key(self):
+        client = Client(TEST_API_KEY, personal_api_key=TEST_API_KEY)
+        with freeze_time("2020-01-01T12:01:00.0000Z"):
+            self.assertRaises(APIError, client.load_feature_flags)
+
+    @mock.patch("posthog.client.decide")
+    @mock.patch("posthog.client.get")
+    def test_feature_enabled_simple(self, patch_get, patch_decide):
+        client = Client(TEST_API_KEY)
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": True,
+                "rollout_percentage": 100,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": 100,
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertTrue(client.feature_enabled("beta-feature", "distinct_id"))
+        self.assertEqual(patch_decide.call_count, 0)
+
+    @mock.patch("posthog.client.decide")
+    @mock.patch("posthog.client.get")
+    def test_feature_enabled_simple_is_false(self, patch_get, patch_decide):
+        client = Client(TEST_API_KEY)
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": True,
+                "rollout_percentage": 0,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": 0,
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertFalse(client.feature_enabled("beta-feature", "distinct_id"))
+        self.assertEqual(patch_decide.call_count, 0)
+
+    @mock.patch("posthog.client.decide")
+    @mock.patch("posthog.client.get")
+    def test_feature_enabled_simple_is_true_when_rollout_is_undefined(self, patch_get, patch_decide):
+        client = Client(TEST_API_KEY)
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": True,
+                "rollout_percentage": None,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": None,
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertTrue(client.feature_enabled("beta-feature", "distinct_id"))
+        self.assertEqual(patch_decide.call_count, 0)
+
+    @mock.patch("posthog.client.get")
+    def test_feature_enabled_simple_with_project_api_key(self, patch_get):
+        client = Client(project_api_key=TEST_API_KEY, on_error=self.set_fail)
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": True,
+                "rollout_percentage": 100,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": 100,
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertTrue(client.feature_enabled("beta-feature", "distinct_id"))
+
+    @mock.patch("posthog.client.decide")
+    def test_feature_enabled_request(self, patch_decide):
+        patch_decide.return_value = {"featureFlags": {"beta-feature": True}}
+        client = Client(TEST_API_KEY, personal_api_key="test")
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": False,
+                "rollout_percentage": 100,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": 100,
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertTrue(client.feature_enabled("beta-feature", "distinct_id"))
+        self.assertEqual(patch_decide.call_count, 0)
+
+    @mock.patch("posthog.client.decide")
+    def test_feature_enabled_request_multi_variate(self, patch_decide):
+        patch_decide.return_value = {"featureFlags": {"beta-feature": "variant-1"}}
+        client = Client(TEST_API_KEY, personal_api_key="test")
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": False,
+                "rollout_percentage": 100,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": 100,
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertTrue(client.feature_enabled("beta-feature", "distinct_id"))
+        # decide not called because this can be evaluated locally
+        self.assertEqual(patch_decide.call_count, 0)
+
+    @mock.patch("posthog.client.get")
+    def test_feature_enabled_simple_without_rollout_percentage(self, patch_get):
+        client = Client(TEST_API_KEY)
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": True,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                        }
+                    ]
+                },
+            }
+        ]
+        self.assertTrue(client.feature_enabled("beta-feature", "distinct_id"))
+
+    @mock.patch("posthog.client.decide")
+    def test_get_feature_flag(self, patch_decide):
+        patch_decide.return_value = {"featureFlags": {"beta-feature": "variant-1"}}
+        client = Client(TEST_API_KEY, personal_api_key="test")
+        client.feature_flags = [
+            {
+                "id": 1,
+                "name": "Beta Feature",
+                "key": "beta-feature",
+                "is_simple_flag": False,
+                "rollout_percentage": 100,
+                "filters": {
+                    "groups": [
+                        {
+                            "properties": [],
+                            "rollout_percentage": 100,
+                        }
+                    ],
+                    "multivariate": {
+                        "variants": [
+                            {"key": "variant-1", "rollout_percentage": 50},
+                            {"key": "variant-2", "rollout_percentage": 50},
+                        ]
+                    },
+                },
+            }
+        ]
+        self.assertEqual(client.get_feature_flag("beta-feature", "distinct_id"), "variant-1")
+        # decide not called because this can be evaluated locally
+        self.assertEqual(patch_decide.call_count, 0)
+
+    @mock.patch("posthog.client.Poller")
+    @mock.patch("posthog.client.decide")
+    def test_feature_enabled_doesnt_exist(self, patch_decide, patch_poll):
+        patch_decide.return_value = {"featureFlags": {}}
+        client = Client(TEST_API_KEY)
+        client.feature_flags = []
+
+        self.assertFalse(client.feature_enabled("doesnt-exist", "distinct_id"))
+        self.assertTrue(client.feature_enabled("doesnt-exist", "distinct_id", True))
+
+    @mock.patch("posthog.client.Poller")
+    @mock.patch("posthog.client.decide")
+    def test_personal_api_key_doesnt_exist(self, patch_decide, patch_poll):
+        client = Client(TEST_API_KEY, personal_api_key="test")
+        client.feature_flags = []
+
+        patch_decide.return_value = {"featureFlags": {"feature-flag": True}}
+
+        self.assertTrue(client.feature_enabled("feature-flag", "distinct_id"))
+
+    @mock.patch("posthog.client.Poller")
+    @mock.patch("posthog.client.get")
+    def test_load_feature_flags_error(self, patch_get, patch_poll):
+        def raise_effect():
+            raise Exception("http exception")
+
+        patch_get.return_value.raiseError.side_effect = raise_effect
+        client = Client(TEST_API_KEY, personal_api_key="test")
+        client.feature_flags = []
+
+        self.assertFalse(client.feature_enabled("doesnt-exist", "distinct_id"))
 
 
 class TestConsistency(unittest.TestCase):
